feat: Add deck-template-preview module
This commit adds the "deck-template-preview" module to the project. This module is responsible for providing a preview of the deck template. It will allow users to see how the template will look before applying it.

The module is exported from the "./deck-template-preview" file.

This change addresses the need for a preview functionality in the deck template feature.